Jeck - [Requires JGE rebuild] Extensive cache improvements.  * Numerous cache fixes, reduced filesystem access.  * Cache fails very gracefully.  * Cache is now a templated class, with individual caches per data-type.  * Much easier to extend.  * Extensively debugged. Try compiling with -DDEBUG_CACHE.  * Caches limits can be set on a per-item basis.  * hgeParticleSystemInfo are now cached, mana particles now fall back to defaults.  * Samples are not cached, but track filesystem misses using the cache backbone.  * Avatars are cached. Default baka avatar is now baka.jpg, to prevent collision with player.
A note on the retrieval types: 
  RETRIEVE_MANAGE puts a resource into a seperate, managed resource list.
    Managed resources are guarenteed valid for the lifetime of the program. 
    Retrieving a managed quad promotes the associated texture to managed. Don't do that by mistake.
    Calls to Resources.Refresh() will attempt to reload managed resources in place.

  RETRIVE_LOCK (and by extension, RETRIEVE_VRAM), returns a resource after locking it.
    A resource may have many locks, and remains in cache until they are all released. 
    If the resource is managed, it returns it unmodified.

A note on quads:
Unlike all other RetrieveWhatever() functions, the default behavior for RetrieveQuad is RETRIEVE_LOCK. Worse, Release(JQuad*) is slow, and will /always/ release one lock from the associated texture.   

There's a long and complicated explanation for this, involving support for live relinking of textures to existing quads, but basically what it means is that we only use RetrieveQuad for quads we intend to store and later Release(). If a temporary quad is needed, the preferred method is to use NEW JQuad* and SAFE_DELETE with RetrieveTexture(). RetrieveTempQuad is also provided, but is only guaranteed until the next call to the cache.

Note that RetrieveCard has none of these problems.