Psyringe - added test suite scripts for i169 - i175, Added several non-bug scripts. Replaced symbiotic_wurm.txt - it now tests token generation after the wurm has been killed. Removed description mentioning Akron Legionnaires which had nothing to do with the rest of the script. Renamed volcanic_island.txt to volcanic_eruption.txt since the script actually has nothing to do with Volcanic Island, updated in-script description as well. Corrected erroneous "endinterrupt" command in three scripts to "endinterruption".
(I actually spent half an hour fiddling with my first interruption script, because the tutorial on the blog doesn't mention the any command to end interrupt sequences. Then I checked the "Counterspell" script, was happy to find the "endinterrupt" keyword, and fiddled for another 30 minutes with it before I found out that this keyword is actually wrong and the parser doesn't understand it at all, it just so happens that the Counterspell script (and two others) work nevertheless because in these scripts the interruption doesn't need to end. I wholeheartedly agree with the passage in wololo's mail in which he says that we could use a "real" test suite tutorial / reference.