corrected typo in lurking evil. just for future notice this card should be moved to crappy. whenever an ability has blah blah:effect that means, {cost}:effect if you can not do it with supported cost in engine, do not add it as {0}:effect && effect this is a workaround we stopped using a very very long time ago. the ability of losing life is not treated the same as "paying" life. same as any other cost.
the issue arose several times while i have added support, and due to a workaround, my support would be flawed, at first i would try to munipulate what i coded to allow for the workarounds to continue working. 
that was the past, i long since stopped doing that, if a card is a workaround, and it breaks, i remove it completely.